pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eden"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "paho-mqtt>=2.0",
    "boto3>=1.34",
    "strands-agents>=0.1",
    "strands-agents-tools>=0.1",
    "mcp>=1.0",
    "requests>=2.31",
    "python-dotenv>=1.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sse-starlette>=2.0",
    "bedrock-agentcore>=1.4.0",
    "structlog>=24.1",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.setuptools.packages.find]
include = ["eden*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "e2e: end-to-end tests hitting real AWS services (Bedrock, DynamoDB)",
]
